Move cycle checking logic inside the dag definition
