+++ airflow-core/src/airflow/models/dag.py
@@ -96,7 +96,6 @@ from airflow.timetables.simple import (
     OnceTimetable,
 )
 from airflow.utils.context import Context
-from airflow.utils.dag_cycle_tester import check_cycle
 from airflow.utils.log.logging_mixin import LoggingMixin
 from airflow.utils.session import NEW_SESSION, provide_session
 from airflow.utils.sqlalchemy import UtcDateTime, lock_rows, with_row_locks
@@ -1489,16 +1488,6 @@ class DAG(TaskSDKDag, LoggingMixin):
             count = 0
             print("Cancelled, nothing was cleared.")
         return count
-
-    def cli(self):
-        """Exposes a CLI specific to this DAG."""
-        check_cycle(self)
-
-        from airflow.cli import cli_parser
-
-        parser = cli_parser.get_parser(dag_parser=True)
-        args = parser.parse_args()
-        args.func(args, self)
 
     @provide_session
     def create_dagrun(
